transfered changes vanna made into the total budget components I created
